Write documentation, add labels, fix checked attribute
We can now add and edit labels to icons, with the input following other
events in real time.

The "checked" attribute of HTML checkboxes requires some special
handling, which is now implemented (so you can just feed it true or false).

Also update toolbox to allow clicks to the div besides the label to
change tool.